# ecommerce/docs/swagger.yaml
definitions:
  controllers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.PasswordInput:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  controllers.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.dataInput:
    properties:
      alamat:
        type: string
      nama:
        type: string
      no_hp:
        type: string
    type: object
  controllers.paymentInput:
    properties:
      provider:
        type: string
    type: object
...
  models.Payment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      provider:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nama:
        type: string
      no_hp:
        type: string
      password:
        type: string
      remember_token:
        type: string
      role:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /payment:
    get:
      description: Get a list of payment.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Payment'
            type: array
      summary: Get all payments.
      tags:
      - Payment
    post:
      description: Creating a new Payment.
      parameters:
      - description: the body to create a new Payment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.paymentInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
      security:
      - BearerToken: []
      summary: Create New Payments.
      tags:
      - Payment
  /payment/{id}:
    delete:
      description: Delete a Payment by id.
      parameters:
      - description: payment id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Payments.
      tags:
      - Payment
    put:
      description: Update Payment by id.
      parameters:
      - description: Payment id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Payment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.paymentInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
      security:
      - BearerToken: []
      summary: Update Payments.
      tags:
      - Payment
...
  /user:
    get:
      description: Get a list of User. (Only Admin Role 1)
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - BearerToken: []
      summary: Get all Users.
      tags:
      - User
  /user/{username}:
    delete:
      description: Delete a User by username. (Admin Role Only)
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one User.
      tags:
      - User
    get:
      description: Get a User by userName.
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerToken: []
      summary: Get a User.
      tags:
      - User
    put:
      description: Only user with the same token or only them can change their own
        password
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: the body to update Password
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.PasswordInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerToken: []
      summary: Update Change Current User Password
      tags:
      - User
  /user/login:
    post:
      description: Logging in to get jwt token to access admin or user api by roles.
      parameters:
      - description: the body to login a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login as as user.
      tags:
      - User
  /user/profil/{username}:
    put:
      description: Only user with the same token or only them can change their own
        Data
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: the body to update Data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.dataInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerToken: []
      summary: Update Change Current User Data
      tags:
      - User
  /user/register:
    post:
      description: registering a user from public access. Default Role values is 0
        (customer); for admin is Role = 1
      parameters:
      - description: the body to register a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register a user.
      tags:
      - User
...
swagger: "2.0"